Replace Node with networkx + pluralize parents
Using networkx is simpler and easier than implementing a custom Node class. The
original impetous was using a directed acyclic graph representation instead of
a tree because Reorganization causes some nodes to have more than one parent.

Speaking of which, the record `parent_key` field has been replaced with the
`parent_keys` field (a list) to allow for more than one parent for a node.

A related change is the `refounder` field in the directory has been made blank
if the member is not a refounder, since the string "No" evaluates as True when
it ideally should evaluate as False.