Move color handling to FamilyTree class
Although this means that not all node attributes can be generated from
MemberRecords alone, this really was never true to begin with (since family
information depends on multiple MemberRecords anyway). This ensures that colors
and family---tree properties, not individual properties---are kept and dealt
with in the family tree class.

Also, the `family` field in MemberRecords was not initialized in __init__,
which made me uncomfortable.