Remove special handling of chapter nodes
All references for chapter nodes have been removed, except in some
comments that need to be reexamined and updaed. I decided it was not
worth all the effort just for one node. I plan to include Kappa Delta
Chapter's node instead as a custom addition included in the settings
file.