Add stage between file reading and tree creation
Previously, the program---for all intents and purposes---directly added
members to the tree at the same time as reading. Now, the members are
read into an intermediate data structure first (`Directory`), which is
then turned into the FamilyTree structure. Hopefully this will allow the
code to be cleaner and allow connecting to SQL without too much trouble.

Note: Affiliations have been removed temporarily.