Allow attributes in settings to be None
All attributes dicts should be dicts in the settings. But if this is a
line in YAML:

  key:
    attributes:

Then key['attributes'] will be None. The new code turns this None into
an empty dict. The new code also provides an empty `attributes` field if
none exists, for custom nodes and edges in settings.