Make colors dependent on component, not root node
The result is that you can give the color for any member of a family,
and that color is then used for the whole family (as determined by weak
connectivity). Previously, you could only give the color to a root of
the family, only the root's descendents would get it, and probably shit
would get fucked if there were a cycle.

(This also makes it easier to add support for trees whose founder members,
having no real bigs, are bigs to each other.)