Init jury-rigged separation of ΔA stuff from rest
The goal is to make a datree.py executable (perhaps with a different
name), with options to load a CSV or SQL. Later, there should be a basic
executable for CSVs without keys, one for CSVs with keys, and one for
DOT. Or something like that. The only Sigma Nu specific stuff should be
in the directory schema (for general Sigma Nu use) and the executable
(for ΔA stuff).