Replace cerberus with voluptuous in row validation
Yeah, I know, I changed my mind /again/. Here's the reasoning: Settings
files are small and nested, which is what I like cerberus for.
Voluptuous is not as elegant with nested structures, IMO, but that
doesn't matter because rows in a table are not really nested.

This brought the runtime of Directory.set_members from around 800 ms in
the last commit to about 230 ms (and compared with about 500 ms in older
commits). Note that affiliations, once again, have been temporarily
removed, so the final speed will be slower.