Remove checks for cycle
This saves barely enough time to make it worth it. It is also a little
redundant: All bigs must have semesters equal to or less than their littles,
which eliminates all cycles except cycles on the same semester (which might be
useful if founders, for example, are bigs to each other). I suppose this
/might/ be an issue if semester checking was removed, but I don't expect that
to be a problem.