Remove unnecessary TODO
The TODO was meant to throw and error if the brother's badge was not in
his affiliations. But this is automatically done in the appropriate
__init__. And even if the badge (external_id) in CiviCRM was not one of
the badges in the CiviCRM relationships, this is to be expected because
the CiviCRM affiliation relationships are optional for primary
affiliations (mainly because I can't expect anyone to update it manually
and because I have not made a way to do it reliably automatic).