Remove explicit defaults in CLI, added some types
They could be confusing because defaults are mostly handled in
validation, not directly in CLI itself. In any case, the click package
was already going to send them to the same values anyway. The types
were added to options that no long had clear types when the default was
taken out.