Modularize readers / turn readers into plugins
A side effect is that this fixes a bug where readers.sql, readers.csv,
etc. weren't working correctly. (This bug was missed in the tests due to
how tests import, but caught when ./snutree.py was actually run. It was
introduced earlier and I caught it only a couple commits ago.)