Separate logging configuration from API
Logging options are now controlled through the cli module (i.e., the
module with the user interface), though most logging code is in the
snutree.logging module so it can share with the qt module in the future.
This allows more control over logging for instances where neither cli
nor qt are required.