Fix affiliation chapter designation assignment
The Affiliation.str_to_designation function was only called when the
Affiliation constructor received a single string (as opposed to a string
and an int), allowing for malformed designations to appear when the
two-argument version of the constructor was used. Moved the
str_to_designation called to the end of the constructor (after the
if-block) and added a new test to catch this problem in the future.