Add examples folder
Includes copies of the sigmanu and chapter trees from tests. Added
basic, custom, and keyed examples. (The custom example is based on the
custom test, but with named and labeled classes.) Includes a simple
example generation script to create/regenerate more CSVs (although the
CSVs are committed because some have had post-generation modifications
to them).