Use UTF-8 for all output and input
Output: Previously it was the system's default encoding. It is now
always UTF-8 (with built-in text-based writers).

Input: Config and input files were already expected to be in UTF-8.
However, passing '-' as a command-line argument to use stdin as input
caused the program to unexpectedly use the system encoding (see the
relevant comment added in this commit). That has been fixed: Input must
now be converted to UTF-8 before being piped to snutree's stdin.