Delegate input stream decoding to reader modules
The first argument of get_table() in all readers must now be interpreted
as a byte stream, not a text stream (whose encoding was already chosen).
(The main argument parser now just opens the files in binary mode, not
text mode with encoding.)

In order to facilitate this, the FileType subclass of argparse.FileType
now handles the case where a binary stream is requested.