Shuffle node order within each component
Previously: The weakly connected components were collected, shuffled,
and for each component the nodes for that component were printed in
sorted order.

Now: The components are still shuffled, but now their members are also
shuffled and printed in random order. (In the old way, the nodes within
each component would have been always sorted in the same order no matter
what the seed.)

This makes changing the seed have much more of an effect, giving us a
little more control to change the layout if it's too messy.